Auth

+++ src/main/java/com/example/demo/controllers/CancionController.java
@@ -4,6 +4,7 @@ import com.example.demo.exceptions.StandardRequestException;
 import com.example.demo.dto.CancionDTO;
 import com.example.demo.models.Cancion;
 import com.example.demo.services.CancionService;
+import com.example.demo.services.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
@@ -13,8 +14,13 @@ import java.util.List;
 @RestController
 public class CancionController {
 
+    private final String AUTHORIZATION_HEADER = "Authorization";
+
     @Autowired
     private CancionService cancionService;
+
+    @Autowired
+    private JwtService jwtService;
 
     @GetMapping(value = "/canciones", produces = MediaType.APPLICATION_JSON_VALUE)
     @CrossOrigin(origins = "http://localhost:4200")
@@ -41,19 +47,25 @@ public class CancionController {
 
     @PostMapping(value = "/cancion", produces = MediaType.APPLICATION_JSON_VALUE)
     @CrossOrigin(origins = "http://localhost:4200")
-    public CancionDTO createCancion(@RequestBody Cancion cancion) throws StandardRequestException {
+    public CancionDTO createCancion(
+            @RequestHeader(name = AUTHORIZATION_HEADER) String token, @RequestBody Cancion cancion) throws StandardRequestException {
+        jwtService.requiresAdmin(token);
         return cancionService.createCancion(cancion);
     }
 
     @PutMapping(value = "/cancion/{cid}", produces = MediaType.APPLICATION_JSON_VALUE)
     @CrossOrigin(origins = "http://localhost:4200")
-    public CancionDTO updateCancion(@PathVariable Long cid, @RequestBody Cancion cancion) throws StandardRequestException {
+    public CancionDTO updateCancion(
+            @RequestHeader(name = AUTHORIZATION_HEADER) String token, @PathVariable Long cid, @RequestBody Cancion cancion) throws StandardRequestException {
+        jwtService.requiresAdmin(token);
         return cancionService.updateCancion(cid, cancion);
     }
 
     @DeleteMapping(value = "/cancion/{cid}", produces = MediaType.APPLICATION_JSON_VALUE)
     @CrossOrigin(origins = "http://localhost:4200")
-    public CancionDTO deleteCancion(@PathVariable Long cid) throws StandardRequestException {
+    public CancionDTO deleteCancion(
+            @RequestHeader(name = AUTHORIZATION_HEADER) String token, @PathVariable Long cid) throws StandardRequestException {
+        jwtService.requiresAdmin(token);
         return cancionService.deleteCancion(cid);
     }
 }
